refactor(test-util): improve TopologyAssert
Documents and updates the `TopologyAssert` utility class. This adds new
methods such as `hasExpectedReplicasCount`, `hasLeaderForEachPartition`,
and `containsBroker`. It removes `isHealthy`, which was previously only
used to check that partitions with a leader were healthy, and instead
merges that check with `hasLeaderForEachPartition`, which would fail if
the leader is not healthy. Finally, it adds a check for the expected
replicas count in `isComplete`.

Additionally, this class was not previously tested, and the assertions
with more complex logic (`isComplete`, `hasLeaderForEachPartition`, and
`hasExpectedReplicasCount`) are now tested.