refactor(qa): fix flakiness on startup of ClusterRule
The `ClusterRule` uses the topology to figure out if the cluster is
ready before a test starts. There was a flaky condition where it might
start when all partitions are present, but before they are healthy. This
change was pulled from #9502 - however in that PR, it was only done in
the broker module's `EmbeddedBrokerRule`.

At the same time, the rule is updated such that everywhere it was
waiting for the topology, now also uses `TopologyAssert`.