docs(CONTRIBUTING): containerize distribution
I always have to search my terminals history for the docker build
command to containerize the zeebe distribution. This made me think that
its probably not clear to other users how to containerize the distro.

This commit adds a short how-to for building the docker image to the
contributing guide as it is part of building Zeebe from source code.