refactor: rm side effects queue usage
Due to our move to event-sourcing we no longer need the side effects, because commands are only processed on processing once and on replay events are applied, which means we don't need to distingush between processing / reprocessing or avoid things like sending responses on reprocessing. Responses can be send immediately on processing, without further delay.