fix(broker): do not log error if follower fails to take snapshot when log is not up-to-date
Alternative options where to prevent taking snapshot in such cases. But it is very difficult to identify this case. All options prevented taking snapshots in scenarios where we want to take snapshot.
1. One option was to return -1 as processed position in ReplayStateMachine when no events after snapshot position is replayed. But this also prevents taking snapshot when there are no new records, but the exporter position increases.
2. Another options was to skip snapshot in StateController, when the determined snapshot position <= position in latest snapshot. This would not fix the bug, because if the exporter position has changed, it will still try to take snapshot and enter the same situation where it cannot find the indexed entry. Other options were attempted, but all failed one or the other edge case.

So, the easiest solution is to treat it as a special case in AsyncSnapshotDirector.

(cherry picked from commit 372367d30f93b62d9c94de8ce58272650bb507fb)