feat: check if blockbuilders equals or contains specific blockbuilder
To generate an execution path where we want to make sure it contains the block we'd want to start at there is 2 things we should know:

1. The block we want to start at
2. Whether or not this block is part of the current branch of the process

This commit takes care of number 2. By adding some logic in each block we can figure out whether the starting block is either the current block, or of it is one of the nested blocks of this block. If this is the case we return true, and we know our execution path must pass this block!