feat: reject command when failing to subscribe
If the engine is unable to subscribe to the events of a Create Process
Instance command with start instructions, then we need to reject the
command. We also don't want the process instance and any of its element
instances to exist, but subscribing to the events happens after creating
the process instance and creating the scopes of the elements to start
at.

To deal with this we need to rollback the transaction, but that's not
possible until we have full transaction support in the engine.

As a workaround we can simply throw an exception during the processing.
The ProcessingStateMachine will handle the exception, rollback the
state changes, won't write the records to the log, blacklist the process
instance (which now doesn't exist), reject the command, and write an
error record.

This workaround does a bit too much: it blacklists the process instance
that wasn't created and writes an error record. It also logs the
exception's stacktrace as an error. Lastly, it rejects the command with
a rather technical error messages. All of these are acceptable for the
situation.