feat: find random execution path when starting at a random block
Trying to find an execution path is a complex task. In the previous commits we've made sure we have the required data we need (the start block and whether this block or one of it's nested blocks equals this start block). The next step is making sure all of the block builders keep the start block into consideration, and guarantee a path is found that contains this block.

For this a check has been introduced in all the block builders. This check will check whether or not this block builder is on the right track towards the start block. It will do this by checking whether or not this block is the start block, or if it contains the start block. If this is the case the builder can start adding steps, or finding an execution path through it's nested blocks.
If it is not the case, we can ignore this builder completely and no steps have to be taken.

We also want to make sure that any blocks coming after our start block will generate execution steps, as we do need the process to fully complete. For this the ExecutionPathContext has been introduced. In this context we can keep track on whether or not we have found the starting block. Since the blocks are traversed in order (to keep the order of the execution steps correct) we know that once we've added the steps of the starting block, we will also need the steps of all the blocks that come after this. This is handled by the foundBlockBuilder flag on the context.

Parallel gateways are a bit of an exception to the normal flow. Instead of only adding the steps for the branch that contains the starting block, we also need to start at the other branches, otherwise the process will get stuck at the joining gateway. To resolve this the ParallelGatewayBlockBuilder will first find a random path for the branch that does contain the starting block. This results in the foundBlockBuilder flag being set to true. Because of this all the other branches will also add their steps. Since we also need to make sure we start the process in these branches we will also add a secondaryStartBlockBuilder to the context. When sending the command we can look into this list and add a start instruction for each of these. For now this starting element will always be the first block of this branch!