feat: start random process anywhere
Please note that this is the first solution I tried. This is not the final solution!

When generating an execution path we should randomly decide if we want to start at a random point, or at the regular start event. If we decide to start anywhere we'll have to find a specific block we'd want to start at. From there can generate the execution path from that block onwards.
We should also send a start command containing the element id of the start block as a start instruction.