fix: equal chance for starting in nested block
Before we would traverse the blocks, each time generating a random boolean to decide if we'd want to go deeper into the nested blocks. This random boolean has a 50% chance of being true. This means when we have a process like this:

start -> task1 -> subprocess(start -> task2 -> end) -> end

The possible starting places here are task1, task2 and the subprocess. First we'd where we want to start in the root block. This would give us a 50/50 chance of rolling task1 or subprocess. If we roll subprocess we will have a 50/50 chance of starting into the nested blocks (task2) or at the subprocess itself. This means statistically that the chances to start at each element is:
- task1: 50%
- subprocess: 25%
- task2: 25%
This is not very well distributed and becomes even worse with more blocks.

With this change we collect a list of all the blocks we could possibly start at. From this block we will randomly pick one. This gives each block an equal chance of being picked to start at.