refactor: dont use restricted record identifier
In Java `record` is a restricted identifier which should not be used as
a variable name. We can replace it with recordValue, because the
ProcessInstanceCreationRecord is the concrete implementation of the
ProcessInstanceCreationRecordValue interface.

Switching to the interface also means we program against the interface
instead of the concrete implementation, which is generally considered
better if possible.

Switching does mean we need to use a different method to retrieve the
start instructions.