refactor: simplify condition
If we are inside this generate method that means one of 3 things:
- This block is the start block
- This block contains the start block
- The start block has already been found

If there is no branch containing the start block, this means the either this block is the start block, or that the start block has already been found. In both cases we should append a step, so just checking if branchContainingStartBlock is empty is sufficient here.