refactor: extract path generation from if-else
The path generation is the same for both scenarios, all it requires is a context that is properly initialized.

We also don't have to explicitly say that we have already found the start block for the default start scenario. The check whether the block should generate an execution path already does this for us.