refactor: combine start elements into a single list
Step 3 of making the context unaware of the existence of block builders.
We no longer need a separation between different start elements. We can put them all in a single list and check if the list contains a specific element id to decide if we need to execute the block or not.