perf: dont copy all values to determine size
The ProcessEngineMetrics needs to know whether the created process
instance had any start instructions. As implementation it used to use
the `getStartInstructions` method to retrieve a list and then it could
check whether this list is empty.

Since the previous commit, the `getStartInstruction` method iterates
over all the entries and makes a copy of each object. This is a rather
costly operation.

To speed up this check (which is called for every created process
instance), we can directly check the size of the ArrayValue inside the
ArrayProperty.

This commit also changes the type of the recordValue method parameter
from the interface to the concrete implementation. The reason for this
is that we need the `hasStartInstructions` method, but this method
cannot be added to the interface. If it would be added to the interface,
it is considered another property in much of the generated code
(immutables/jackson).