refactor(engine): move command sender into SideEffectContext
(will not compile)

In this step the command sender is moved into the SideEffectContext. The assumption is that if this is a side effect, it should only be accessed when side effects are produced.

Doing this refactoring exposes some places where a command sernder is used outside of side effects