deps(parent): promotes JNA as production library for native bindings
Moves the JNA dependency up with the normal dependencies and not simply
for convergence. Additionally, we add a new `add-opens` directive to
surefire to make sure we can reflectively access FileDescriptor#fd for
native calls. While in production it's enough to add the right manifest
entry in the zeebe-util JAR, when running the tests via IntelliJ, these
are not picked up so reflection always fails. It's simpler to add it for
the whole project (though I'm happy if we find a workaround).