fix: don't update state in tests
Updating the state in tests can result in unexpected error in RocksDb. The reason for this is that the transaction is used by the StreamProcessor. When we modify the state we are using the same transaction, but this time from the test thread, resulting in 2 threads accessing the same transaction (we reuse the same transaction everytime by resetting it).

There is a very low chance that these two threads are interfering with each other. To circumvent the test could create it's own context on the ZeebeDB when it needs to update the state. In this case it is a lot easier to just use a made-up value for the key of a record, as it doesn't matter what we use.