feat(util): add FileUtil#preallocate
Adds a method to preallocate a new file of the expected length,
effectively reserving the disk space for later use. The current
implementation is a "dumb" one, which simply writes 4KB blocks of 0s
until the expected length is reached (meaning the size may be up to 4KB
- 1 byte greater than the desired length, which is acceptable for our
  use case).

While potentially slow, this is not in a hot path, and will be later
optimized with a native syscall to fallocate.