test(util): add tests for FileUtil#preallocate
Adds basic tests for `FileUtil#preallocate`. A new dependency was added,
`jnr-posix`, which allows us to check the actual size on disk of the
file in UNIX systems. This gives us the accurate number of blocks
reserved for the file, on disk, even on file systems with compression or
sparse files (i.e. most modern Linux systems), and also ensures that we
don't only read the metadata (e.g. what `Files.size` returns) but really
guarantee we reserved the disk space.