refactor: make `DiskUsageMonitor` available in partition context
Before this change, adding and removing a disk space usage listener
was implemented directly by the broker context and the partition manager
only had a callback to add usage listeners.

Since we now want to add partition scoped components as disk space usage
listeners directly, the `DiskSpaceMonitor` is made available in the
`PartitionTransitionContext`.