feat(broker): compact the logs based on oldest available snapshots
For taking backups asynchronously, we have to keep the snapshot needed for the backup until the backup is completed. We should also not compact the log segments that are needed for the backup. To enable this we will now only compact the logs based on the oldest snapshot. The backup manager is expected to reserve the snapshot so that the snapshot store doesn't delete the snapshot. Thus even if there is a new snapshot, we won't compact the logs until the backup is completed.