fix(broker): throw recoverable exception when LogStorageSteps fails due to newer transitions
Since the role transitions in raft and zeebe partitions are asychronous, it is possible that while zeebe partition is installing leader services the raft has already transitioned to follower role. In this case, LogStorageStep fails because it cannot get the appender as the raft is not in the leader role anymore. Previously this fails the installation and the partition goes to an unhealthy state. But since this is an expected error, we could ignore it and continue with the installation of the new role. To enable this, instead of throwing IllegalStateException, we now throw a recoverable exception which won't be considered as a fatal error by ZeebePartition.