feat(journal): preallocate segment files
Allows preallocating segment files if configured to do so. On segment
creation, we now preallocate the segment file by default before using
it. If it already existed (but was unused by the log so far), we take
the easy way out and simply delete/recreate it instead of attempting to
grow/shrink it.