test(journal): test segment preallocation
Add a new dependency on `jnr-posix` to accurately get the size of the
file on disk. This is necessary to avoid tests passing if we only change
the file's metadata - for example, when mmap-ing a file with a mapping
of length X, then the file will report it has a length of X, even though
it has possibly only one block allocated on disk.