refactor(scheduler): ensure futures are completed on failure
When explicitly failing an actor, we want to ensure that anyone waiting
on its start and/or close futures is notified that these will never
complete. To avoid unnecessary retry loops, the close future is
completed successfully, and the start future is completed exceptionally.
This is because a caller may want to react differently if the actor did
not start correctly, but most likely does not care if it did not close
correctly.

(cherry picked from commit f25e231fadf6c38e1a3ef5c99d9859bfd828a8c0)