refactor(engine): move StreamProcessorContext
Moves StreamProcessorContext into ..zeebe.streamprocessors package as this is the context object for the platform part.
Adds a new RecordProcessorContext interface to capture the information that is needed by record processors. This is to later thin out StreamProcessorContext.

For now, however StreamProcessorContext simply implements that interface to keep the changes minimal.

Once the engine abstraction is established, each time can continue refactoring their halfs of the interface