refactor(engine): simplify ReadonlyStreamProcessorContext
Previously, this interface was designed by following the rule "This interface has read only methods for everything in StreamProcessorContext". This is now changed to "This interface has everything that is actually needed." The switch was made, because StreamProcessorContext has grown quite large and we want to reduce its size. With this change, we can better see, which information is actually needed by clients of the interface.

Also, we were able to remove some stuff from StreamProcessorContext which was not used