refactor(engine): remove actor from RecordProcessorContext
The purpose of this commit is to introduce a clear interface and isolate the engine from the scheduling package.

Some code is kept in the productive interface to support existing tests.

Repeated tasks (runAtFixedRate) are replaced by tasks that reschedule themselves. The changes might have subtle effects on the behavior of the scheduled tasks. This needs to be carefully considered during review.

This commit does not change direct access to log stream. Part of the engine abstraction PoC is that scheduled tasks can return a list of follow-up records. This part is out of scope for now.