test(engine): fix tests
So this is a strange one. Without these changes, the test
TimerStartEventTest#shouldCreateTimerFromFeelExpression failed.
In the log there was an ArithmeticException when calculating the nanos.
The test used a date way in the future, and it turns out we will run
into this issue whenever we schedule a task for >290 years in the future.

So I did a comparison with the original code. The original code also did
produce that exception, but for some reason the exception was ignored silently.
No log no nothing. The code just went on.

The test only checks for the entry in the state, and there it is fine.

So the bottom line is that there is a subtle difference between the old code
and the new code in the way exceptions are propagated.

I did some time boxed research, but could not pinpoint the exact difference.

So in the end, I decided to rewrite the test to a more reasonable date.