refactor: allow sonar to highlight missing case
By moving the runtime exception throwing out of the switch case
statement (and removing the default case), we allow sonar to create a
code quality issue out of a missing case. So if we add a new Intent, we
see a code quality issue at compile time, and don't have to wait for the
exception at runtime.