refactor(engine): move LastProcessedPositionState to stream processor
The last processed position state is only used by the StreamProcessor. Therefore, it
was moved into the streamprocessor package. The existing classes and interfaces were
moved as is. There is probably a way to simplify this code, but this is something we
can still do after the refactoring.

In the tests, minimal effort was undertaken to make the tests pass. The test classes
are very complex and ripe for an overhaul, but this would have increased the change
set considerably