refactor(engine): StreamProcessorContext no longer implements TypedRecordProcessorContext
This is to separate the two contexts. One is used for the stream processor, the other in the engine.

This is an intermediate step where the engine upon initializing
returns objects that are still needed in the stream processor
context. SOme will disappear once the engine abstraction
refactoring is complete. Others may require additional refactoring
later on.