fix: increment the number of taken sequence flows when starting a process anywhere
In the scenario where we start a process instance at a Parallel Gateway we will have to increment the number of taken sequence flows manually. If we don't do this the new state transition guard will find no taken sequence flows, where it has N amount of incoming flows. Resulting in a command rejection.
By incrementing the number of taken sequence flows for each incoming flow of the parallel gateway we can circumvent this rejection and activate the gateway as expected.