refactor: remove can activate check
This check has been moved to the ProcessInstanceStateTransitionGuard. We can always send the activate command. If a parallel gateway cannot be activated the command gets rejected.
This might have a small impact on performance, as now we are sending commands and rejecting them, as opposed to not sending the commands at all.