refactor: remove can activate check
This check has been moved to the ProcessInstanceStateTransitionGuard. We can always send the activate command. If a parallel gateway cannot be activated the command gets rejected.
This might have a small impact on performance, as now we are sending commands and rejecting them, as opposed to not sending the commands at all.

(cherry picked from commit 016c280396350b7dbfb8bf87f672de3b822e8e74)