feat: implement ProcessInstanceModificationRecord
This provides an initial implementation of the process instance
modification record.

The main things to note is that the getters for the ArrayProperties have
been written for safety over performance. They copy each element as
collecting the elements in a single pass results in a list with the same
size, but each element is the last element in the original array.

Performance is not that important for the modification command, as it is
an advanced feature that should only be used sporadically. However, if
an additional speed boost is needed, the implementation provides hasX
methods. These allow the caller to check whether there are entries in
the ArrayProperty, without having to use the slow getter.