feat(broker): support modification command in API
To be fair, this commit is not necessary to support the modification
command in the broker's command api. As the record can be read and
deserialized from the bytes, however this speeds up the deserialization
because the record is already available as a Java object (which is a
costly object to create).