feat(elasicsearch): support modification records
Thhis sets up the elasticsearch exporter to support exporting of the
PROCESS_INSTANCE_MODIFICATION records, but does not yet actually export
them. The ES exporter is configured to NOT export these records for now.
If we find a need to export these records, we can simply flip the
configuration flag.

To support the exporting of these records, this commmit:

- adds the template for the PROCESS_INSTANCE_MODIFICATION records so
  they can be exported to elastic search.
- adds a mapping between the value type and this template
- provides a way to enable exporting these records, even though it is
  disabled by default for now.

All properties of the record are covered by the template, but the
variables are disabled. This is similar to the PROCESS_INSTANCE_CREATION
records which also don't export the variables they contain. Only the
variables themselves are not exported, the variable instructions are
exported to show that the modification record did contain some variable
instructions. Note that any variable changes are still exported via the
VARIABLE records.