feat(backup): add processors for checkpoint records
The implementation follows the new RecordProcessor api. It is not quite clear what will be the "init" method. So that is kept blank for now. My assumption is that some of the initialization in the constructor will be moved to the init methods, because the dependencies needed to initialize them are not available when the object is created.

Since there is only one processor and one event applier, I have hard-coded the mapping between the Intent and the processor/event applier. We can also re-use the RecordProcessorMap and TypedRecordProcessor from the engine. But to keep the dependency to the engine minimal, they are not used. Once the streamprocessor/engine api refactoring is done, we could revisit them and see if we want to reuse some of the "framework" implemented by the engine to handle multiple records types. Until then let's keep this simple.