style(journal): removes Journal prefix from class
We use the prefix `JournalSegment` for all classes related to a single
segment. It's fairly obvious that a segment is a journal segment in this
case, so no need to be so verbose. Instead, all classes with
`JournalSegment` are renamed to `Segment` only. Similarly, we drop the
`Mapped` prefix which is historical from when we had multiple
reader/writer strategies.