refactor(journal): moves packages to better reflect structure of module
The root package includes the journal API and any common classes used
across the journal.

The `record` package includes classes which deal with record
serialization. These are not inherently file specific, as they can be
serialized in-memory as well.

The `file` classes deal with the persistence layer of the journal.