refactor(engine): move retry logic out of processing result
This reverts an earlier change. The reason is that during the
earlier refactoring the information about the record position
got lost. This caused some failing tests.

The logic was moved back to ProcessingStateMachine. The current
implementation is a bit weird, because it will be implementation-
dependant when the position is actually being updated.