refactor(journal): use strategy pattern instead of boolean state
Replaces the use of a boolean field with a segment allocation strategy.
This may seem YAGNI at the moment, but it was very useful to test
different ways of pre-allocating, which is still not 100% out of the
question. As I'd already written it, I figured we could keep just this
even if we discard the other strategies for now.