refactor(engine): treat sending response as special side effect
Before, the command was sent as part of the side effects.
This is no longer ideal. Among others, we want to have a reset
for follow up records, response and side effects.

With the old code, the reset would not unset the sideEffectProducer
managed in ProcessingStateMachine.

By making the sending of the response a special case, we also
make sure that a reset is independent of the side effect producer.

With the new code, the side effect producer can now be null,