fix(engine): fix failing tests
The tests failed for two reasons:
The process instance key was not set.
Some intent JobIntent.COMPLETE which should not be blacklisted at all, but was

This was because the old code:
zeebeState.getBlackListState().tryToBlacklist(typedCommand, errorRecord::setProcessInstanceKey);

Did check the intents and did set process instance key with this
strange setter.

Now, we don't use the logic anymore and instead create en error event.
However, the event applier is not aware of the logic and has
no access to the intent of the record that caused the error.