refactor(engine): introduce smaller interface for TypedCommandWriter
This commit is a bit larger than I would like. It could be split up into a small commit and a large commit,
so I just left it as one large commit.

It started because I wanted to implement a TypedCommandWriter that uses ProcessingResultBuilder as backend
instead of accessing the writers directly.

Turns out, TypedCommandWriter contains reset() and flush() methods which will no longer be allowed.
(There will be a reset() method for the full result, but not for individual parts of the result).

So I created a smaller interface. RestrictedTypedCommandWriter which does not contain these two methods.
(will be renamed in next commit)

Then I went through the current uses of TypedCommandWriter and replaced all with RestrictedTypedCommandWriter.

In one place this was not possible: DeploymentDistributionBehavior, which schedules a background task that
uses reset and flush.

For this use case, a new class LegacyTask was introduced. We need to migrate all background tasks anyway.
The idea of LegacyTask is to be a bag to capture all unwanted dependencies.

Then there was a need to schedule such a legacy task and provide the dependencies when the task is run.

The effect of both of these are:
- Only the background task has a dependency to a writer with flush and reset; and this task is fully in control by the platform
- All other places in the engine only write, but don't use reset or flush