feat: allow inter-partition commands with given record key
To use inter-partition command sending for deployment distribution
without refactoring deployment distribution, commands need to be written
with a provided record key. This is somewhat unfortunate, and we'd
prefer not supporting this natively in the inter-partition command
sender and receiver, but we had to make a practical choice.

The method is marked as deprecated and contains a hint that this is
only necessary to support deployment distribution. We should remove it
again as soon as possible. The field in the sbe message is marked as
optional and set to the canonical null value when the sender does not
provide a record key.