fix(engine): fix failing tests
For some unknown reason we need to set the hasResponse flag back to true.
This should not be necessary, but it is. Setting it to true by default eliminates
~168 failed tests. Since the direct access shall be replaced with a buffered one
I think there is no point in investigating this further.

Finally, the error induced in StreamProcessorHealthTest must be moved to
a different method to reflect the way the code flow works now.