fix(broker): accept case insensitive exporter configuration
Fixes a bug where deserialization of exporter configuration was case
sensitive, such that a property `NUMBEROFSHARDS` would not map to an
instance field `numberOfShards`.

This isn't perfect, as now if we have two fields with the same name but
different cases, only one of them will be set, using the latest value
in the map (where the order is undefined). It should solve the most common
cases however.

(cherry picked from commit aaab69417622d6592f6903676350ec0e7b3a7886)