test: redistributing dmn deployment fails
A critical bug was found related to redistributing DMN resources. When
the distribution isn't acknowledged, it is retried until it is. This
retry also occurs on the restart of the deployment partition, if the
deployment distribution hasn't been completed for that partition yet.

This retry results in an additional distribute deployment command on the
other partition. Once processed it writes the DISTRIBUTED event which
when applied inserts the decision and drg into the state. The second
time the event is written again and applied again, in which case the
insert fails due to consistency checks of the insert method.

This test doesn't check the implementation of the insert (using the
insert method), but checks the behavior. If the insert fails then it
also doesn't write the DISTRIBUTED event, as the exception will make
sure the entire transaction is rolled back. This test verifies that the
DISTRIBUTED event is correctly written to the log and so without errors.
It actually checks that this event is written twice, after the command
was written twice. After this we can be sure that the retried
distribution of the dmn resource in the deployment was succesfully
processed.

Note that this solution differs from the alternative: rejecting the
retried DISTRIBUTE command if the resources are already known. However,
this solution is more complex and not suitable for patching a critical
bug. We can have a look at that solution in future versions.

Co-authored-by: Nico korthout <nico.korthout@camunda.com>
(cherry picked from commit 9fd942c399c7222891b60b4060b0b72d5bd4ac58)