fix: backoff for retrying deployment distribution
Since we have changed the behavior slightly and now only acknowledge
the distribution after processing the command on the receiver partition,
we need to make sure that we don't overload slow partitions by retrying
distribution too frequently.

Here we keep track of the retry count to implement a simple exponential
backoff, statically configured to start of at 10 seconds until it
reaches a maximum of 5 minutes, doubling every time.

Since we keep the retry counter in-memory, the backoff is reset when
the deployment partition is recovered. Whether that is good behavior
depends on the specific reason why a distribution is pending. For
network issues this might be beneficial as a new leader for the
deployment partition might not have the same network issues. If the
receiving partition is unavailable or lagging, restarting with no
backoff is bad but acceptable.