test(qa): wait until message is published before restarting the broker
The test was flaky - it sometime failed in 'correlate message after restart' while waiting for the instance to be completed. It is not clear why exactly the test fails, but a probable reason is that the message was not published before the broker was restarted.

(cherry picked from commit fcef5b6dfc078e411b96021800e359ca0367cffa)