test(qa): increase timeout
The test fails because the topology is not updated when the message subscription command is sent. So it sent the command to the old leader. The command will be retried only after 10s. So we increase the request timeout in the test.

The topology will be eventually updated. If there is a bug in the topology listener and the leader is never updated, the test will fail even after the increased timeout.

(cherry picked from commit e1db654054ba3245d003f26161214b1d575da418)