feat(broker): add a request handler to process take backup request
The handler receives TAKE_BACKUP request, writes and new record to the logstream. The handler will not sent the response back to the gateway. The response will be sent by the processor after processing this record. This way the gateway knows that the request was processed and there is no need to retry.

We have also added a BackupResponse which is unused now. This will used by the other backup related commands which will be added later. For those requests, the response will be sent by this handler. For that BackupApiResponseWriter must be extended to allow returning a response.