feat(clients/java): introduce new ModifyProcessInstanceCommand
This commits adds support for sending a ModifyProcessInstanceCommand in the java client. The interface lets you add activate/terminate instructions for elements. When activating an element it also provides the option to add variables.

To keep a clear distinction between the element that is currently being activated / terminated it is required to call the and() method when switching to a new element.