feat(broker): unsubscribe handler when closing the handler
The lifecycle of the handler is per partition leader role. When the broker steps down as leader, if there is no handler the gateway gets a "NoRemoteMessageHandler" exception as the error. This messsage is not useful and would be difficult to debug. Hence we add a additional handler that returns and error message indicating it is not the leader.