fix: add variable instruction to activate instruction
GRPC objects are immutable. Once the activate instruction is build we can't just add a variable instruction to it. Instead we will have to remove the activate instruction from the request builder. After this we can copy the activate instruction and add the new variable instruction by using the toBuilder() method. This method will create a builder of the instruction and set all the attributes that it already contained.
Finally we have to add this new activate instruction to the request builder.