fix(engine): reuse key in PI modification event
The process instance modification event's key depends on the incoming
(user) data.

If the related command already has a key set, then it should be used as
the key of the event as well. That way it's more clear that they refer
to the same entity.

If the key is not yet set (i.e. -1), then we should take the process
instance key from the record value as the entity that this command
referred to, and so we need to write the event as well for that key.

Note that this situation makes it clear that the key of the command may
differ from the process instance key specified in its value. We could
consider to tighten this in the gateway, only allowing to write the
command with -1 or with the same key as specified by the process
instance key property.